Align deflateParams() and its documentation in zlib.h.
This updates the documentation to reflect the behavior of
deflateParams() when it is not able to compress all of the input
data provided so far due to insufficient output space.  It also
assures that data provided is compressed before the parameter
changes, even if at the beginning of the stream.

Change-Id: Icd23381fb36ee11ff7000d41ce5cc8c9b6b63cac
Signed-off-by: Alex Naidis <alex.naidis@linux.com>